change buttons style
